requirements.txt
json
pickle
random
re
warnings
gensim.models
matplotlib.pyplot
numpy
pandas
seaborn
nltk.tokenize
sklearn.decomposition
sklearn.ensemble
sklearn.manifold
sklearn.metrics
sklearn.model_selection
sklearn.neighbors
sklearn.neural_network